% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bib.R
\name{read_bib}
\alias{read_bib}
\title{Import bib files and clean BibTeX references}
\usage{
read_bib(
  path = ".",
  tags = c("bibtexkey", "category", "author", "title", "year", "volume", "booktitle",
    "editor", "journal", "pages", "institution", "publisher", "doi", "url", "abstract",
    "keywords", "annote"),
  categories = NULL,
  pattern = NULL
)
}
\arguments{
\item{path}{a character of length 1. The name of the folder containing
\code{.bib} files to read. All \code{.bib} files found in this folder will be
imported.}

\item{tags}{a vector of characters. BibTeX fields to extract
(e.g. \code{"author"}, \code{"year"}, \code{"title"}, etc.). See \strong{Details} section.}

\item{categories}{(optional) a vector of characters. Publications categories
to import (e.g. \code{"article"}, \code{"book"}, etc.). See \strong{Details} section.
Default is \code{NULL} (all publications will be retrieved).}

\item{pattern}{(optional) a vector of characters. This argument is used to
select \code{.bib} files. For instance, if all \code{.bib} files start with
\strong{CESAB-} (i.e. \verb{CESAB-***.bib}), user can add \code{pattern = 'CESAB-'} to
only read these \code{.bib} files.}
}
\value{
A \code{data.frame} with references fields in columns and references in
rows.
}
\description{
Imports \code{.bib} files and cleans BibTeX references using the function
\code{\link[bib2df:bib2df]{bib2df::bib2df()}}.
}
\details{
\strong{Valid \code{tags} values}:
\itemize{
\item \code{category}: type of publication
\item \code{bibtexkey}: unique identifier of the publication
\item \code{annote}: user annotations
\item \code{author}: authors names
\item \code{booktitle}: title of the book
\item \code{chapter}: chapter number (for book chapter)
\item \code{crossref}: Crossref identifier
\item \code{edition}: edition number
\item \code{editor}: editors names
\item \code{institution}: institution name (for thesis and report)
\item \code{journal}: journal name (for scientific article)
\item \code{month}: month of the date of publication
\item \code{number}: issue number
\item \code{pages}: pages range or pages number or article number
\item \code{publisher}: publisher name (i.e. Elsevier, etc.)
\item \code{series}: series name
\item \code{title}: title of the publication
\item \code{volume}: volume number
\item \code{year}: year of the date of publication
\item \code{doi}: DOI (without https://doi.org/)
\item \code{abstract}: abstract of the publications
\item \code{keywords}: authors keywords
\item \code{url}: URL of the only publication
\item \code{isbn}: ISSN identifier of the book/journal
\item \code{pmid}: PMID identifier
}

\strong{Valid \code{categories} values}:
\itemize{
\item \code{article}: scientific paper
\item \code{book}: book
\item \code{incollection}: book chapter
\item \code{inproceedings}: proceeding article
\item \code{phdthesis}: PhD thesis
\item \code{techreport}: technical report
}
}
\examples{
path_to_bibs <- system.file("extdata", package = "rbibtools")
refs <- read_bib(path = path_to_bibs)
}
