% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bibtex.R
\name{read_bibtex}
\alias{read_bibtex}
\title{Read BiBTeX references exported by Mendeley}
\usage{
read_bibtex(
  path = ".",
  tags = c("bibtexkey", "category", "author", "title", "year", "volume", "booktitle",
    "editor", "journal", "pages", "institution", "publisher", "doi", "url", "abstract",
    "keywords", "annote"),
  categories = NULL,
  filename = NULL,
  pattern = NULL
)
}
\arguments{
\item{path}{the folder containing BiBTeX files to read.}

\item{tags}{a vector of characters. BiBTeX fields to extract
(e.g. author, year, etc.). See below \strong{Details} section.}

\item{categories}{(optional) a vector of characters. Publications categories
to import (e.g. article, book, etc.). See below \strong{Details} section.
Default is \code{NULL} (all publications will be retrieved).}

\item{filename}{the name of the xlsx file to create.}

\item{pattern}{(optional) a character used to select BiBTeX files. For
instance, if all BiBTeX start with \strong{CESAB-} (i.e. \verb{CESAB-***.bib}) user
can add \code{pattern = 'CESAB-'} to only read these BiBTeX files.}
}
\value{
A data frame with BiBTeX fields in columns and references in rows.
}
\description{
Reads BiBTeX references exported by Mendeley (or not) and stores metadata
into a data frame using the function \code{\link[bib2df:bib2df]{bib2df::bib2df()}}.
}
\details{
\strong{Valid tags values}:
\itemize{
\item \code{category}: type of publication
\item \code{bibtexkey}: unique identifier of the publication
\item \code{annote}: user annotations
\item \code{author}: authors names
\item \code{booktitle}: title of the book
\item \code{chapter}: chapter number (for book chapter)
\item \code{crossref}: Crossref identifier
\item \code{edition}: edition number
\item \code{editor}: editors names
\item \code{institution}: institution name (for thesis and report)
\item \code{journal}: journal name (for scientific article)
\item \code{month}: month of the date of publication
\item \code{number}: issue number
\item \code{pages}: pages range or pages number or article number
\item \code{publisher}: publisher name (i.e. Elsevier, etc.)
\item \code{series}: series name
\item \code{title}: title of the publication
\item \code{volume}: volume number
\item \code{year}: year of the date of publication
\item \code{doi}: DOI (without https://doi.org/)
\item \code{abstract}: abstract of the publications
\item \code{keywords}: authors keywords
\item \code{url}: URL of the only publication
\item \code{isbn}: ISSN identifier of the book/journal
\item \code{pmid}: PMID identifier
}

\strong{Valid categories values}:
\itemize{
\item \code{article}: scientific paper
\item \code{book}: book
\item \code{incollection}: book chapter
\item \code{inproceedings}: proceeding article
\item \code{phdthesis}: PhD thesis
\item \code{techreport}: technical report
}
}
\examples{
\dontrun{
refs <- read_bibtex(path = ".", filename = "~/Desktop/references", 
                    pattern = "CESAB-")
}
}
