% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmendeley-package.R
\docType{package}
\name{rmendeley-package}
\alias{rmendeley}
\alias{rmendeley-package}
\title{rmendeley: Read BiBTeX references exported by Mendeley}
\description{
Reads and cleans BiBTeX references exported by 
    Mendeley (or others) and exports cleaned references as an Excel file.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FRBCesab/rmendeley}
  \item Report bugs at \url{https://github.com/FRBCesab/rmendeley/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Casajus \email{nicolas.casajus@fondationbiodiversite.fr} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID})

}
\keyword{internal}
