% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_bib.R
\name{write_bib}
\alias{write_bib}
\title{Create a bib file from a data.frame}
\usage{
write_bib(data, file = "", append = FALSE)
}
\arguments{
\item{data}{a \code{data.frame} with references fields in columns and references
in rows.}

\item{file}{a \code{character} of length 1. The name of \code{.bib} file to create.}

\item{append}{a \code{logical}. If \code{TRUE} the \code{data.frame} will be appended to
an existing file. Default is \code{FALSE}.}
}
\value{
No returned value.
}
\description{
Create a bib file from a data.frame
}
\examples{
## Import .bib file(s) ----
path_to_bibs <- system.file("extdata", package = "rbibtools")
refs <- rbibtools::read_bib(path = path_to_bibs)

## Export data.frame to .bib files ----
filename <- tempfile()
rbibtools::write_bib(refs, filename)
}
