% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbibtools-package.R
\docType{package}
\name{rbibtools-package}
\alias{rbibtools}
\alias{rbibtools-package}
\title{rbibtools: Toolbox to Handle BibTeX Files}
\description{
Imports, cleans, and exports BibTeX references exported by references management softwares (e.g. Mendeley, Zotero).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FRBCesab/rbibtools}
  \item Report bugs at \url{https://github.com/FRBCesab/rbibtools/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Casajus \email{rdev.nc@gmail.com} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID}) [copyright holder]

}
\keyword{internal}
