% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmendeley-package.R
\docType{package}
\name{rmendeley-package}
\alias{rmendeley}
\alias{rmendeley-package}
\title{rmendeley: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several 
    lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FRBCesab/rmendeley}
  \item Report bugs at \url{https://github.com/FRBCesab/rmendeley/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Casajus \email{nicolas.casajus@fondationbiodiversite.fr} (\href{https://orcid.org/0000-0002-5537-5294}{ORCID})

}
\keyword{internal}
